subscriber/app/requirements.txt
google-cloud-pubsub==2.8.0
pandas==1.3.4
tabulate==0.8.9